# docs/swagger.yaml
basePath: /v1
...
host: localhost:3010
info:
  contact:
    email: rhanysuf24@gmail.com
    name: Ehancup
    url: http://www.swagger.io/support
  description: This is an API for learning.
  termsOfService: http://swagger.io/terms/
  title: Learn Golang gin-gonic REST API
  version: "1.0"
paths:
...
  /upload/single:
    post:
      consumes:
      - multipart/form-data
      description: Upload Single File with formData
      parameters:
      - description: file will be uploaded
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Upload Single File
      tags:
      - upload
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"